+++ services/blockchain-indexer/shared/dataService/business/auth/index.js
@@ -14,7 +14,9 @@
  *
  */
 const { getAuthAccountInfo } = require('./accountInfo');
+const { getAuthMultiSigRegMsgSchema } = require('./multiSigRegMsgSchema');
 
 module.exports = {
 	getAuthAccountInfo,
+	getAuthMultiSigRegMsgSchema,
 };
